Correcting Collation Mapping Query for Varbinary.

+++ v2/sourcedb-to-spanner/src/test/java/com/google/cloud/teleport/v2/source/reader/io/jdbc/iowrapper/JdbcIoWrapperTest.java
@@ -43,7 +43,7 @@ import com.google.cloud.teleport.v2.source.reader.io.schema.SourceTableSchema;
 import com.google.cloud.teleport.v2.spanner.migrations.schema.SourceColumnType;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import java.math.BigInteger;
+import java.math.BigDecimal;
 import java.sql.SQLException;
 import org.apache.beam.sdk.io.jdbc.JdbcIO;
 import org.apache.beam.sdk.transforms.PTransform;
@@ -467,7 +467,7 @@ public class JdbcIoWrapperTest {
                     .setCardinality(42L)
                     .setIsUnique(true)
                     .build()))
-        .isEqualTo(BigInteger.class);
+        .isEqualTo(BigDecimal.class);
     assertThrows(
         SuitableIndexNotFoundException.class,
         () ->
